requirements.txt
asgiref==3.7.2
cfgv==3.4.0
commitizen==3.9.0
distlib==0.3.7
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
filelock==3.12.4
gunicorn==20.1.0
identify==2.5.29
inflection==0.5.1
nodeenv==1.8.0
packaging==23.1
Pillow==9.5.0
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.5
